pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiosmtpd>=1.4.6",
    "cryptography>=44.0.2",
    "dkimpy>=1.1.8",
    "dnspython>=2.7.0",
    "email-validator>=2.2.0",
    "flask-login>=0.6.3",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "marshmallow>=3.26.1",
    "psycopg2-binary>=2.9.10",
    "pyjwt>=2.10.1",
    "sqlalchemy>=2.0.39",
    "werkzeug>=3.1.3",
    "redis>=5.2.1",
]
